+++ enlighten/ui/HelpFeature.py
@@ -56,12 +56,17 @@ class HelpFeature(EnlightenFeature):
             Ctrl-G enter Gain
             Ctrl-H toggle between Hardware and scope
             Ctrl-L toggle Laser firing
+            Ctrl-M toggle library Matching
             Ctrl-N enter new Note
             Ctrl-P Pause/Play
+            Ctrl-Q Quit
             Ctrl-R take/clear Reference
             Ctrl-S Save measurement
             Ctrl-T enter integration Time
             Ctrl-X toggle eXpert mode
+
+            Ctrl-Shift-D toggle DALAI-RAMAN
+            Ctrl-Shift-S export all Saved measurements
 
             Ctrl-* Auto-Raman
             Ctrl-% enter laser power
@@ -69,9 +74,22 @@ class HelpFeature(EnlightenFeature):
             Ctrl-^ connect currently-selected plugin
             Ctrl-, Prev Spectrometer
             Ctrl-. Next Spectrometer
-            
-            Ctrl-Up/Down step values
+
+            Ctrl-Up/Down scroll palette/hardware
             Ctrl-Left/Right move Scope cursor"""))
+
+        """
+        Unused:
+            ctrl-I
+            ctrl-J
+            ctrl-K
+            ctrl-O
+            ctrl-U
+            ctrl-V
+            ctrl-W
+            ctrl-Y
+            ctrl-Z
+        """
 
     def help_callback(self):
         url = self.HELP_URL